% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosCategoriasEconomicas.R
\name{getDadosCategoriasEconomicas}
\alias{getDadosCategoriasEconomicas}
\title{get Dados Categorias Economicas}
\usage{
getDadosCategoriasEconomicas(path = "data/orcamento/",
  filename = "categorias_economicas.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Categorias Economicas
}
\description{
Retorna Informacoes de Categorias Economicas da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosCategoriasEconomicas(): Retorna data.frame com dados de Categorias Economicas. O Data.frame conterá uma amostra aleatória da população
}
