% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosCodigoReceitaNomenclaturaTCE.R
\name{getDadosCodigoReceitaNomenclaturaTCE}
\alias{getDadosCodigoReceitaNomenclaturaTCE}
\title{get Codigo Receita Nomenclatura}
\usage{
getDadosCodigoReceitaNomenclaturaTCE(path = "data/orcamento/",
  filename = "categorias_economicas.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Codigo Receita Nomenclatura do TCE
}
\description{
Retorna Informacoes de Codigo Receita Nomenclatura TCE da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosCodigoReceitaNomenclaturaTCE(): Retorna data.frame com dados de Nomenclatura de Receita do Tribunal de Contas.
O Data.frame conterá uma amostra aleatória da população
}
