% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosConvenios.R
\name{getDadosConvenios}
\alias{getDadosConvenios}
\title{get Dados Convenios}
\usage{
getDadosConvenios(path = "data/convenios/",
  filename = "siconv_convenio.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Convenios do Sistema de Convenios (Siconv)
}
\description{
Retorna Informacoes de Convenios da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosConvenios(): Retorna data.frame com dados de Convênios. O Data.frame conterá uma amostra aleatória da população
}
