% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosDespesas.R
\name{getDadosDespesas}
\alias{getDadosDespesas}
\title{get Dados Despesas}
\usage{
getDadosDespesas(path = "data/orcamento/",
  filename = c("despesa_lei131_2009_random.csv",
  "despesa_lei131_2010_random.csv", "despesa_lei131_2011_random.csv",
  "despesa_lei131_2012_random.csv", "despesa_lei131_2013_random.csv",
  "despesa_lei131_2014_random.csv", "despesa_lei131_2015_random.csv",
  "despesa_lei131_2016_random.csv", "despesa_lei131_2017_random.csv",
  "despesa_lei131_2018_random.csv"))
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Despesas. Agrega os anos 2009 a 2018
}
\description{
Retorna Informacoes de Despesas da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosDespesas(): Retorna data.frame com dados de Despesas. O Data.frame conterá uma amostra aleatória da população
}
