% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosGuiaReceitaDespesaExtra.R
\name{getDadosGuiaReceitaDespesaExtra}
\alias{getDadosGuiaReceitaDespesaExtra}
\title{get Dados Guia Receitas Despesas}
\usage{
getDadosGuiaReceitaDespesaExtra(path = "data/orcamento/",
  filename = "guia_receita_despesa_extra_random.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Receitas e Despesas Extras
}
\description{
Retorna Informacoes de Receitas e Despesas Extras da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosGuiaReceitaDespesaExtra(): Retorna data.frame com dados de Receitas e Despesas Extras. O Data.frame conterá uma amostra aleatória da população
}
