% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosItensEmpenho.R
\name{getDadosItensEmpenho}
\alias{getDadosItensEmpenho}
\title{get Dados Itens Empenho}
\usage{
getDadosItensEmpenho(path = "data/orcamento/",
  filename = c("itens_empenho_2009_random.csv",
  "itens_empenho_2010_random.csv", "itens_empenho_2011_random.csv",
  "itens_empenho_2012_random.csv", "itens_empenho_2013_random.csv",
  "itens_empenho_2014_random.csv", "itens_empenho_2015_random.csv",
  "itens_empenho_2016_random.csv", "itens_empenho_2017_random.csv",
  "itens_empenho_2018_random.csv"))
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados Itens Empenho
}
\description{
Retorna Informacoes de Itens Empenho da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosItensEmpenho(): Retorna data.frame com dados de Itens Empenho. O Data.frame conterá uma amostra aleatória da população
}
