% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosLicitacoes.R
\name{getDadosLicitacoes}
\alias{getDadosLicitacoes}
\title{get Dados Licitacoes}
\usage{
getDadosLicitacoes(path = "data/licitacoes/",
  filename = "licitacoes_random.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Licitacoes
}
\description{
Retorna Informacoes de Publicidade da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosLicitacoes(): Retorna data.frame com dados de Licitacoes. O Data.frame conterá uma amostra aleatória da população
}
