% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosNaturezaDespesas.R
\name{getDadosNaturezaDespesas}
\alias{getDadosNaturezaDespesas}
\title{Dados Natureza Despesas}
\usage{
getDadosNaturezaDespesas(path = "data/orcamento/",
  filename = "grupo_natureza_despesa.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Natureza das Despesas
}
\description{
Retorna Informacoes de Natureza das Despesas da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosNaturezaDespesas(): Retorna data.frame com dados de Natureza das Despesas. O Data.frame conterá uma amostra aleatória da população
}
