% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosPessoal.R
\name{getDadosPessoal}
\alias{getDadosPessoal}
\title{get Dados Pessoal}
\usage{
getDadosPessoal(path = "data/pessoal/", filename = "pessoal_random.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Pessoal
}
\description{
Retorna Informacoes de Pessoal da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosPessoal(): Retorna data.frame com dados de Publicidade. O Data.frame conterá uma amostra aleatória da população
}
