% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosPropostasConvenios.R
\name{getDadosPropostasConvenios}
\alias{getDadosPropostasConvenios}
\title{get Dados Propostas Convenios}
\usage{
getDadosPropostasConvenios(path = "data/convenios/",
  filename = "siconv_proposta.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Propostas de Convenios do Sistema de Convenios (Siconv)
}
\description{
Retorna Informacoes de Propostas de Convenios da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosPropostasConvenios(): Retorna data.frame com dados de Propostas de Convênios. O Data.frame conterá uma amostra aleatória da população
}
