% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosPublicidade.R
\name{getDadosPublicidade}
\alias{getDadosPublicidade}
\title{get Dados Publicidade}
\usage{
getDadosPublicidade(path = "data/publicidade/",
  filename = "publicidade_random.csv")
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Publicidade
}
\description{
Retorna Informacoes de Publicidade da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosPublicidade(): Retorna data.frame com dados de Publicidade. O Data.frame conterá uma amostra aleatória da população
}
