% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDadosReceitas.R
\name{getDadosReceitas}
\alias{getDadosReceitas}
\title{get Dados Receitas}
\usage{
getDadosReceitas(path = "data/orcamento/",
  filename = c("receita_lei131_2009_random.csv",
  "receita_lei131_2010_random.csv", "receita_lei131_2011_random.csv",
  "receita_lei131_2012_random.csv", "receita_lei131_2013_random.csv",
  "receita_lei131_2014_random.csv", "receita_lei131_2015_random.csv",
  "receita_lei131_2016_random.csv", "receita_lei131_2017_random.csv",
  "receita_lei131_2018_random.csv"))
}
\arguments{
\item{path}{Caminho para o Arquivo}

\item{filename}{Nome do arquivo com Amostra. Intervalo de Confiança de 99%, Margem de erro de 5%}
}
\value{
Dados de Receitas
}
\description{
Retorna Informacoes de Receitas da Prefeitura de Joao Pessoa
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getDadosReceitas(): Retorna data.frame com dados de Receitas. O Data.frame conterá uma amostra aleatória da população
}
