% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoTabela.R
\name{getInfoTabela}
\alias{getInfoTabela}
\title{get Info Tabela}
\usage{
getInfoTabela(tabela = "pessoal")
}
\arguments{
\item{tabela}{Qual Tabela terá a Informação Recuperada.
Opções: pessoal, publicidade, receitas, despesas, empenho, receitasdespesasextras, convenios, propostas}
}
\value{
Informacoes sobre a tabela
}
\description{
Retorna Informacoes Sobre os Campos dos Dados Abertos
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getInfoTabela(): Retorna Colunas e Tipo de Cada coluna no formato data.frame
}
