% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTabelaDadosQualificados.R
\name{getTabelaDadosQualificados}
\alias{getTabelaDadosQualificados}
\title{get Tabela Dados Qualificados}
\usage{
getTabelaDadosQualificados(obj = NA, tabela = "", categoria = "",
  anomaly_threshold = 0.8)
}
\arguments{
\item{obj}{Objeto contendo dados (Amostra) da Tabela para Avaliação.}

\item{tabela}{Qual Tabela terá a Informação Recuperada.}

\item{categoria}{(Publicidade, Pessoal, Convenio, Orcamento)}

\item{anomaly_threshold}{Threshold de Anomalias
Opções: pessoal, publicidade, receitas, despesas, empenho, receitasdespesasextras, convenios, propostas}
}
\value{
Informacoes de Qualificacao da Tabela passada como parametro
}
\description{
Retorna Informacoes de Qualificacao para Tabela (Somente variaves problematicas)
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getTabelaDadosQualificados(): Retorna Qualificacao dos Dados da Tabela (Somente variaves problematicas)
}
