% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTabelaDadosTotal.R
\name{getTabelaDadosTotal}
\alias{getTabelaDadosTotal}
\title{get Tabelas Dados Total}
\usage{
getTabelaDadosTotal(obj = NA, tabela = "", categoria = "",
  anomaly_threshold = 0.8)
}
\arguments{
\item{obj}{Objeto contendo dados (Amostra) da Tabela para Avaliação.}

\item{tabela}{Qual Tabela terá a Informação Recuperada.}

\item{categoria}{(Publicidade, Pessoal, Convenio, Orcamento)}

\item{anomaly_threshold}{Threshold de Anomalias
Opções: pessoal, publicidade, receitas, despesas, empenho, receitasdespesasextras, convenios, propostas}
}
\value{
Informacoes de Qualificacao da Tabela passada como parametro
}
\description{
Retorna Informacoes de Qualificacao para Tabela (Sumario de Todos os Dados)
Disponivel em: transparencia.joaopessoa.pb.gov.br/download
}
\examples{
getTabelaDadosTotal(): Retorna Qualificacao Completa dos Dados da Tabela (Sumario de Todos os Dados)
}
