% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualificaDadosAbertosJP.R
\name{qualificaDadosAbertosJP}
\alias{qualificaDadosAbertosJP}
\title{qualifica Dados Abertos JP}
\usage{
qualificaDadosAbertosJP(anomaly_threshold = 0.8)
}
\arguments{
\item{data.atualizacao}{Data em que os dados foram obtidos. Formato dd/mm/aaaa}
}
\value{
Dados de Qualificação da Prefeitura de Joao Pessoa
}
\description{
Função para Qualificação de Dados Abertos de Joao Pessoa (Somente os campos problemáticos)
Todos os CVSs utilizados está disponivel em: transparencia.joaopessoa.pb.gov.br/download (10/02/2018)
}
\examples{
qualificaDadosAbertosJP(): Retorna data.frame com painel de Qualificação de Dados Abertos de Joao Pessoa
nas seguintes áreas: Pessoas, Convenios (Propostas e Convenios Firmados), Publicidade,
Orçamentarios (Despesas, Receitas, Empenho, Guia de Receitas e Despesas Extras)
}
